pyproject.toml
[project]
name = "ubuweb-mirror"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "lxml-html-clean>=0.4.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "requests-html>=0.10.0",
    "tqdm>=4.67.1",
    "youtube-dl>=2021.12.17",
]
